+++ externallib.php
@@ -57,6 +57,7 @@ class local_booking_external extends external_api {
         return new external_function_parameters(
             array(
                 'courseid'  => new external_value(PARAM_INT, 'The course id', VALUE_DEFAULT),
+                'filter'  => new external_value(PARAM_RAW, 'The results filter', VALUE_DEFAULT),
             )
         );
     }
@@ -65,16 +66,17 @@ class local_booking_external extends external_api {
      * Retrieve instructor's booking.
      *
      * @param int $courseid The course id for context.
-     * @param int $categoryid The category id for context.
+     * @param string $filter The filter to show students, inactive (including graduates), suspended, and default to active.
      * @return array array of slots created.
      * @throws moodle_exception if user doesnt have the permission to create events.
      */
-    public static function get_bookings_view($courseid) {
+    public static function get_bookings_view(int $courseid, string $filter) {
         global $PAGE, $COURSE;
 
         // Parameter validation.
         $params = self::validate_parameters(self::get_bookings_view_parameters(), array(
                 'courseid' => $courseid,
+                'filter' => $filter,
                 )
             );
 
@@ -86,7 +88,7 @@ class local_booking_external extends external_api {
         if (empty($COURSE->subscriber))
             $COURSE->subscriber = new subscriber($courseid);
 
-        list($data, $template) = get_bookings_view($courseid);
+        list($data, $template) = get_bookings_view($courseid, '', $filter);
 
         return $data;
     }
@@ -297,7 +299,7 @@ class local_booking_external extends external_api {
                 if (core_user::get_user($userid, 'alternatename')->alternatename == $alternatename) {
                     // get engine type integrated data
                     if (subscriber::has_integration('aircraft')) {
-                        $enginetyperec = subscriber::get_integrated_data('aircraft', 'enginetype', $logentry->get_aircraft());
+                        $enginetyperec = subscriber::get_integrated_data('aircraft', 'aircraftinfo', $logentry->get_aircraft());
                         if (!empty($enginetyperec))
                             $logentry->set_enginetype($enginetyperec['engine_type'] == 'single' ? 'SE' : 'ME');
                     }
@@ -862,7 +864,7 @@ class local_booking_external extends external_api {
         if (empty($COURSE->subscriber))
             $COURSE->subscriber = new subscriber($courseid);
 
-        $student = $COURSE->subscriber->get_active_student($USER->id);
+        $student = $COURSE->subscriber->get_student($USER->id);
         $warnings = array();
 
         // add new slots after removing previous ones for the week
@@ -938,7 +940,7 @@ class local_booking_external extends external_api {
         if (empty($COURSE->subscriber))
             $COURSE->subscriber = new subscriber($courseid);
 
-        $student = $COURSE->subscriber->get_active_student($USER->id);
+        $student = $COURSE->subscriber->get_student($USER->id);
         $warnings = array();
 
         // remove all week's slots for the user to avoid updates